requirements.txt
jina
streamlit
pillow

